A pointer was cast to an unsigned int and I changed it to an unsigned long so it would work on 64 bit pointers.

+++ code/casa/implement/Containers/test/tHashMap.cc
@@ -79,7 +79,7 @@ main() {
 	exit(1);
     }
 
-    srand((unsigned int) &x);
+    srand((unsigned long) &x);
     for (int y=0; y < 500; y++) {
 	Int index = rand() % x;
 	if ( index != hash(block[index]) )
